fs/fs-writeback.c: add a new writeback list for sync
wait_sb_inodes() currently does a walk of all inodes in the filesystem to
find dirty one to wait on during sync.  This is highly inefficient and
wastes a lot of CPU when there are lots of clean cached inodes that we
don't need to wait on.

To avoid this "all inode" walk, we need to track inodes that are currently
under writeback that we need to wait for.  We do this by adding inodes to
a writeback list on the sb when the mapping is first tagged as having
pages under writeback.  wait_sb_inodes() can then walk this list of
"inodes under IO" and wait specifically just for the inodes that the
current sync(2) needs to wait for.

Define a couple helpers to add/remove an inode from the writeback list and
call them when the overall mapping is tagged for or cleared from
writeback.  Update wait_sb_inodes() to walk only the inodes under
writeback due to the sync.

With this change, filesystem sync times are significantly reduced for fs'
with largely populated inode caches and otherwise no other work to do. 
For example, on a 16xcpu 2GHz x86-64 server, 10TB XFS filesystem with a
~10m entry inode cache, sync times are reduced from ~7.3s to less than
0.1s when the filesystem is fully clean.

Link: http://lkml.kernel.org/r/1466594593-6757-2-git-send-email-bfoster@redhat.com
Signed-off-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Josef Bacik <jbacik@fb.com>
Signed-off-by: Brian Foster <bfoster@redhat.com>
Reviewed-by: Jan Kara <jack@suse.cz>
Tested-by: Holger Hoffstätte <holger.hoffstaette@applied-asynchrony.com>
Cc: Al Viro <viro@ZenIV.linux.org.uk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>