crc32: use ktime_get_ns() for measurement
The crc32 test function measures the elapsed time in nanoseconds, but uses
'struct timespec' for that.  We want to remove timespec from the kernel
for y2038 compatibility, and ktime_get_ns() also helps make the code
simpler here.

It is also slightly better to use monontonic time, as we are only
interested in the time difference.

Link: http://lkml.kernel.org/r/20160617143932.3289626-1-arnd@arndb.de
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Cc: "David S . Miller" <davem@davemloft.net>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>