uapi: export lirc.h header
This header contains the userspace API for lirc.

This is a fixup for b7be755733dc44c72 ("[media] bz#75751: Move internal
header file lirc.h to uapi/").  It moved the header to the right place,
but it forgot to add it at Kbuild.  So, despite being at uapi, it is not
copied to the right place.

Fixes: b7be755733dc44c72 ("[media] bz#75751: Move internal header file lirc.h to uapi/")
Link: http://lkml.kernel.org/r/320c765d32bfc82c582e336d52ffe1026c73c644.1468439021.git.mchehab@s-opensource.com
Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>
Cc: Alec Leamas <leamas.alec@gmail.com>
Cc: <stable@vger.kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>