mm, hugetlb: fix huge_pte_alloc BUG_ON
Zhong Jiang has reported a BUG_ON from huge_pte_alloc hitting when he runs
his database load with memory online and offline running in parallel.  The
reason is that huge_pmd_share might detect a shared pmd which is currently
migrated and so it has migration pte which is !pte_huge.

There doesn't seem to be any easy way to prevent from the race and in fact
seeing the migration swap entry is not harmful.  Both callers of
huge_pte_alloc are prepared to handle them.  copy_hugetlb_page_range will
copy the swap entry and make it COW if needed.  hugetlb_fault will back
off and so the page fault is retries if the page is still under migration
and waits for its completion in hugetlb_fault.

That means that the BUG_ON is wrong and we should update it.  Let's simply
check that all present ptes are pte_huge instead.

Link: http://lkml.kernel.org/r/20160721074340.GA26398@dhcp22.suse.cz
Signed-off-by: Michal Hocko <mhocko@suse.com>
Reported-by: zhongjiang <zhongjiang@huawei.com>
Acked-by: Naoya Horiguchi <n-horiguchi@ah.jp.nec.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>